+++ NEWS.md
@@ -1,4 +1,14 @@
-# svrep (development version)
+# svrep 0.4.1
+
+* Minor Updates and Bug Fixes:
+  * Fix bug in [#15](https://github.com/bschneidr/svrep/issues/15), where 
+  bootstrap conversion of multistage survey design objects with `as_bootstrap_design()` 
+  would throw an error when user manually specified weights in `svydesign()`.
+  
+  * Creation of Rao-Wu-Yue-Beaumont bootstrap replicate weights is now faster
+  and takes less computer memory.
+
+  * Typo fix in vignettes.
 
 # svrep 0.4.0